backend/requirements.txt
requests
beautifulsoup4
python-dotenv
sqlalchemy
alembic
psycopg2-binary
yt-dlp
